[Refactor] Extract module NameSimilarity from UselessAssignment
